msm: clock: Fix list rate to handle rate > LONG_MAX
While at it, also convert the list_rate return type from int to long to be
consistent with the rest of the ops.

+++ arch/arm/mach-msm/clock-8960.c
@@ -3047,7 +3047,7 @@ static int pix_rdi_clk_reset(struct clk *c, enum clk_reset_action action)
 	return branch_reset(&to_pix_rdi_clk(c)->b, action);
 }
 
-static int pix_rdi_clk_list_rate(struct clk *c, unsigned n)
+static long pix_rdi_clk_list_rate(struct clk *c, unsigned n)
 {
 	if (pix_rdi_mux_map[n])
 		return n;
